delete dtype=int64

+++ pandas/core/base.py
@@ -1154,16 +1154,16 @@ class IndexOpsMixin(object):
         dtype: int64
 
         >>> x.searchsorted(4)
-        array([3], dtype=int64)
+        array([3])
 
         >>> x.searchsorted([0, 4])
-        array([0, 3], dtype=int64)
+        array([0, 3])
 
         >>> x.searchsorted([1, 3], side='left')
-        array([0, 2], dtype=int64)
+        array([0, 2])
 
         >>> x.searchsorted([1, 3], side='right')
-        array([1, 3], dtype=int64)
+        array([1, 3])
 
         >>> x = pd.Categorical(['apple', 'bread', 'bread',
                                 'cheese', 'milk'], ordered=True)
@@ -1171,10 +1171,10 @@ class IndexOpsMixin(object):
         Categories (4, object): [apple < bread < cheese < milk]
 
         >>> x.searchsorted('bread')
-        array([1], dtype=int64)     # Note: an array, not a scalar
+        array([1])     # Note: an array, not a scalar
 
         >>> x.searchsorted(['bread'], side='right')
-        array([3], dtype=int64)
+        array([3])
         """)
 
     @Substitution(klass='IndexOpsMixin')
